+++ packages/result/src/ResultAsync.ts
@@ -603,20 +603,15 @@ export interface ResultAsync<T, E> extends PromiseLike<Result<T, E>> {
      *
      * assert(await OkAsync(Ok(1)).equal(Ok(Ok(1))));
      * assert(await OkAsync(Ok(1)).equal(Ok(OkAsync(1))));
-     * assert(await OkAsync(Ok(1)).equal(Promise.resolve(Ok(Ok(1)))));
-     * assert(await OkAsync(Ok(1)).equal(OkAsync(Promise.resolve(Ok(1)))));
      * assert(await OkAsync(Ok(1)).equal(OkAsync(OkAsync(1))));
-     * assert(await OkAsync(Promise.resolve(Ok(1))).equal(Promise.resolve(Ok(OkAsync(1)))));
      * assert(await OkAsync(OkAsync(1)).equal(OkAsync(Ok(1))));
      *
      * assert((await OkAsync([1]).equal(Ok([1]))) === false);
      * assert((await OkAsync({ foo: 1 }).equal(Promise.resolve(Ok({ foo: 1 })))) === false);
-     * assert((await ErrAsync({ message: 'err' }).equal(ErrAsync({ message: 'err' }))) === false);
+     * assert((await ErrAsync({ msg: 'err' }).equal(ErrAsync({ msg: 'err' }))) === false);
      *
      * assert((await OkAsync(Ok([1])).equal(Ok(Ok([1])))) === false);
      * assert((await OkAsync(Ok([1])).equal(OkAsync(OkAsync([1])))) === false);
-     * assert((await OkAsync(Promise.resolve(Ok([1]))).equal(OkAsync(Ok([1])))) === false);
-     * assert((await OkAsync(Promise.resolve(Ok({ foo: 1 }))).equal(Ok(OkAsync({ foo: 1 })))) === false);
      * assert((await OkAsync(OkAsync({ foo: 1 })).equal(OkAsync(OkAsync({ foo: 1 })))) === false);
      * ```
      */